Fix CI: out-of-bounds read in a new test, ambiguous to_string on libc++
Two CI failures, both mine, both from the 0.12.4 audit branch.

asan-ubsan: AddressSanitizer reported a global-buffer-overflow, a 19-byte read
of an 18-byte string literal, in the media attachment-disposition scenario added
by this branch. The payload was written as one literal with a hand-counted
length -- "\x00\x01\x02\x03binary-payload" is 18 bytes and the count said 19, so
the std::string(char const*, size_t) constructor read one byte past the end.
Rebuilt as an explicit 4-byte NUL prefix plus a NUL-terminated literal, so no
hand-counted total remains to drift. The leading NUL bytes are the point of the
fixture -- an octet-stream body a strlen-based construction would truncate to
nothing -- so they stay.

freebsd/openbsd build-and-test and packages: master_key_file_identity called
std::to_string on file_time_type::rep. libstdc++ makes that `long`, which picks
an overload unambiguously; libc++ uses a type matching none of them exactly, so
every BSD job failed to compile while every Linux job passed. Both counts are
now widened to fixed-width types before formatting.

Verified rather than assumed: reproduced the sanitizer job locally with the same
command CI runs (scripts/build-linux.sh --buildtype debug --sanitize
address,undefined under clang) -- build-asan/meson-logs/testlog.txt reports
Ok 49, Fail 0, with zero AddressSanitizer/UndefinedBehaviorSanitizer reports and
zero failed assertions. The conformance suite, which held the overflow, passes.

Note on how this was missed: the ordinary build reads one byte past a string
literal without complaint, which is exactly the class of bug the sanitizer job
exists to catch. The BSD failure could not have been caught locally at all --
the toolchain difference is the bug.

+++ src/homeserver/auth_service.cpp
@@ -123,7 +123,13 @@ namespace
         {
             return {};
         }
-        return path + '\0' + std::to_string(size) + '\0' + std::to_string(written.time_since_epoch().count());
+        // Both counts are widened to fixed-width types before formatting.
+        // file_time_type::rep is implementation-defined -- libstdc++ makes it
+        // `long`, which picks a std::to_string overload unambiguously, while
+        // libc++ (FreeBSD, OpenBSD) uses a type that matches none of them
+        // exactly and the call fails to compile.
+        return path + '\0' + std::to_string(static_cast<std::uint64_t>(size)) + '\0' +
+               std::to_string(static_cast<std::int64_t>(written.time_since_epoch().count()));
     }
 
     [[nodiscard]] auto token_hmac_keys(HomeserverRuntime const& runtime) -> TokenHmacKeys
@@ -1059,8 +1065,7 @@ auto login_local_user_by_id(HomeserverRuntime& runtime, std::string_view user_id
                                  {"reason",    "account deactivated",  false}
         },
                              observability::LogEventSeverity::warning, observability::AuditCategory::auth,
-                             "login.rejected", std::string{user_id}, std::string{device_id},
-                             "403:account deactivated");
+                             "login.rejected", std::string{user_id}, std::string{device_id}, "403:account deactivated");
         return make_operation_result(false, {}, "account deactivated", 403U);
     }
